+++ src/OpenTK.Windowing.GraphicsLibraryFramework/GLFW.cs
@@ -1227,9 +1227,9 @@ namespace OpenTK.Windowing.GraphicsLibraryFramework
         /// This function sets the value of an attribute of the specified window.
         /// </para>
         /// <para>
-        /// The supported attributes are <see cref="WindowAttributeSetter.Decorated"/>,
-        /// <see cref="WindowAttributeSetter.Resizable"/>, <see cref="WindowAttributeSetter.Floating"/>,
-        /// <see cref="WindowAttributeSetter.AutoIconify"/> and <see cref="WindowAttributeSetter.FocusOnShow"/>.
+        /// The supported attributes are <see cref="WindowAttribute.Decorated"/>,
+        /// <see cref="WindowAttribute.Resizable"/>, <see cref="WindowAttribute.Floating"/>,
+        /// <see cref="WindowAttribute.AutoIconify"/> and <see cref="WindowAttribute.FocusOnShow"/>.
         /// </para>
         /// <para>
         /// Some of these attributes are ignored for full screen windows.
@@ -1245,7 +1245,7 @@ namespace OpenTK.Windowing.GraphicsLibraryFramework
         /// <param name="value"><c>true</c> or <c>false</c>.</param>
         /// <remarks>
         /// <para>
-        /// Calling <see cref="GetWindowAttrib"/> will always return the latest value,
+        /// Calling <see cref="GetWindowAttrib(Window*, WindowAttributeGetBool)"/> will always return the latest value,
         /// even if that value is ignored by the current mode of the window.
         /// </para>
         /// <para>
@@ -1255,7 +1255,7 @@ namespace OpenTK.Windowing.GraphicsLibraryFramework
         /// Possible errors include <see cref="ErrorCode.NotInitialized"/>, <see cref="ErrorCode.InvalidEnum"/>, <see cref="ErrorCode.InvalidValue"/> and <see cref="ErrorCode.PlatformError"/>.
         /// </para>
         /// </remarks>
-        public static unsafe void SetWindowAttrib(Window* window, WindowAttributeSetter attribute, bool value)
+        public static unsafe void SetWindowAttrib(Window* window, WindowAttribute attribute, bool value)
         {
             glfwSetWindowAttrib(window, attribute, value ? GLFW_TRUE : GLFW_FALSE);
         }
@@ -2967,7 +2967,7 @@ namespace OpenTK.Windowing.GraphicsLibraryFramework
         /// <a href="https://www.glfw.org/docs/latest/window_guide.html#window_hints_hard">hard constraints</a>.
         /// This includes the size of the window, especially for full screen windows.
         /// To query the actual attributes of the created window, framebuffer and context,
-        /// see <see cref="GetWindowAttrib" />, <see cref="GetWindowSize" /> and <see cref="GetFramebufferSize" />.
+        /// see <see cref="GetWindowAttrib(Window*, WindowAttributeGetBool)" />, <see cref="GetWindowSize" /> and <see cref="GetFramebufferSize" />.
         /// </para>
         /// <para>
         /// To create a full screen window, you need to specify the monitor the window will cover.
@@ -3101,7 +3101,7 @@ namespace OpenTK.Windowing.GraphicsLibraryFramework
         /// <a href="https://www.glfw.org/docs/latest/window_guide.html#window_hints_hard">hard constraints</a>.
         /// This includes the size of the window, especially for full screen windows.
         /// To query the actual attributes of the created window, framebuffer and context,
-        /// see <see cref="GetWindowAttrib" />, <see cref="GetWindowSize" /> and <see cref="GetFramebufferSize" />.
+        /// see <see cref="GetWindowAttrib(Window*, WindowAttributeGetBool)" />, <see cref="GetWindowSize" /> and <see cref="GetFramebufferSize" />.
         /// </para>
         /// <para>
         /// To create a full screen window, you need to specify the monitor the window will cover.
@@ -3523,9 +3523,45 @@ namespace OpenTK.Windowing.GraphicsLibraryFramework
         /// Possible errors include <see cref="ErrorCode.NotInitialized"/>, <see cref="ErrorCode.InvalidEnum"/> and <see cref="ErrorCode.PlatformError"/>.
         /// </para>
         /// </remarks>
-        public static unsafe bool GetWindowAttrib(Window* window, WindowAttributeGetter attribute)
+        public static unsafe bool GetWindowAttrib(Window* window, WindowAttributeGetBool attribute)
         {
             return glfwGetWindowAttrib(window, attribute) == GLFW_TRUE;
+        }
+
+        /// <inheritdoc cref="GetWindowAttrib(Window*, WindowAttributeGetBool)"/>
+        public static unsafe int GetWindowAttrib(Window* window, WindowAttributeGetInt attribute)
+        {
+            return glfwGetWindowAttrib(window, attribute);
+        }
+
+        /// <inheritdoc cref="GetWindowAttrib(Window*, WindowAttributeGetBool)"/>
+        public static unsafe ClientApi GetWindowAttrib(Window* window, WindowAttributeGetClientApi attribute)
+        {
+            return glfwGetWindowAttrib(window, attribute);
+        }
+
+        /// <inheritdoc cref="GetWindowAttrib(Window*, WindowAttributeGetBool)"/>
+        public static unsafe ContextApi GetWindowAttrib(Window* window, WindowAttributeGetContextApi attribute)
+        {
+            return glfwGetWindowAttrib(window, attribute);
+        }
+
+        /// <inheritdoc cref="GetWindowAttrib(Window*, WindowAttributeGetBool)"/>
+        public static unsafe OpenGlProfile GetWindowAttrib(Window* window, WindowAttributeGetOpenGlProfile attribute)
+        {
+            return glfwGetWindowAttrib(window, attribute);
+        }
+
+        /// <inheritdoc cref="GetWindowAttrib(Window*, WindowAttributeGetBool)"/>
+        public static unsafe ReleaseBehavior GetWindowAttrib(Window* window, WindowAttributeGetReleaseBehavior attribute)
+        {
+            return glfwGetWindowAttrib(window, attribute);
+        }
+
+        /// <inheritdoc cref="GetWindowAttrib(Window*, WindowAttributeGetBool)"/>
+        public static unsafe Robustness GetWindowAttrib(Window* window, WindowAttributeGetRobustness attribute)
+        {
+            return glfwGetWindowAttrib(window, attribute);
         }
 
         /// <summary>
@@ -3801,6 +3837,41 @@ namespace OpenTK.Windowing.GraphicsLibraryFramework
         public static unsafe void MaximizeWindow(Window* window) => glfwMaximizeWindow(window);
 
         /// <summary>
+        /// This function returns the Windows specific window handle (HWND).
+        /// </summary>
+        /// <param name="window">The window to query.</param>
+        /// <returns>The Windows specific window handle (HWND).</returns>
+        public static unsafe IntPtr GetWin32Window(Window* window) => glfwGetWin32Window(window);
+
+        /// <summary>
+        /// This function returns the macos specific window handle (NSWindow).
+        /// </summary>
+        /// <param name="window">The window to query.</param>
+        /// <returns>The macos specific window handle (NSWindow).</returns>
+        public static unsafe IntPtr GetCocoaWindow(Window* window) => glfwGetCocoaWindow(window);
+
+        /// <summary>
+        /// This function returns the x11 specific window handle (Window).
+        /// </summary>
+        /// <param name="window">The window to query.</param>
+        /// <returns>The x11 specific window handle (Window).</returns>
+        public static unsafe uint GetX11Window(Window* window) => glfwGetX11Window(window);
+
+        /// <summary>
+        /// This function returns the glx specific window handle (Window).
+        /// </summary>
+        /// <param name="window">The window to query.</param>
+        /// <returns>The glx specific window handle (Window).</returns>
+        public static unsafe uint GetGLXWindow(Window* window) => glfwGetGLXWindow(window);
+
+        /// <summary>
+        /// This function returns the wayland specific window handle (struct wl_surface*).
+        /// </summary>
+        /// <param name="window">The window to query.</param>
+        /// <returns>The wayland specific window handle (struct wl_surface*).</returns>
+        public static unsafe IntPtr GetWaylandWindow(Window* window) => glfwGetWaylandWindow(window);
+
+        /// <summary>
         /// <para>
         /// This function sets the maximization callback of the specified window,
         /// which is called when the window is maximized or restored.
@@ -3824,7 +3895,7 @@ namespace OpenTK.Windowing.GraphicsLibraryFramework
             Window* window,
             GLFWCallbacks.WindowMaximizeCallback callback)
         {
-            return glfwSetWindowMaximizeCallback(window, Marshal.GetFunctionPointerForDelegate(callback));
+            return glfwSetWindowMaximizeCallback(window, callback);
         }
 
         /// <summary>
@@ -3851,7 +3922,7 @@ namespace OpenTK.Windowing.GraphicsLibraryFramework
             Window* window,
             GLFWCallbacks.FramebufferSizeCallback callback)
         {
-            return glfwSetFramebufferSizeCallback(window, Marshal.GetFunctionPointerForDelegate(callback));
+            return glfwSetFramebufferSizeCallback(window, callback);
         }
 
         /// <summary>
@@ -3878,7 +3949,7 @@ namespace OpenTK.Windowing.GraphicsLibraryFramework
             Window* window,
             GLFWCallbacks.WindowContentScaleCallback callback)
         {
-            return glfwSetWindowContentScaleCallback(window, Marshal.GetFunctionPointerForDelegate(callback));
+            return glfwSetWindowContentScaleCallback(window, callback);
         }
 
         /// <summary>
@@ -3992,7 +4063,7 @@ namespace OpenTK.Windowing.GraphicsLibraryFramework
             Window* window,
             GLFWCallbacks.CharCallback callback)
         {
-            return glfwSetCharCallback(window, Marshal.GetFunctionPointerForDelegate(callback));
+            return glfwSetCharCallback(window, callback);
         }
 
         /// <summary>
@@ -4029,7 +4100,7 @@ namespace OpenTK.Windowing.GraphicsLibraryFramework
             Window* window,
             GLFWCallbacks.CharModsCallback callback)
         {
-            return glfwSetCharCallback(window, Marshal.GetFunctionPointerForDelegate(callback));
+            return glfwSetCharModsCallback(window, callback);
         }
 
         /// <summary>
@@ -4112,7 +4183,7 @@ namespace OpenTK.Windowing.GraphicsLibraryFramework
             Window* window,
             GLFWCallbacks.CursorEnterCallback callback)
         {
-            return glfwSetCursorEnterCallback(window, Marshal.GetFunctionPointerForDelegate(callback));
+            return glfwSetCursorEnterCallback(window, callback);
         }
 
         /// <summary>
@@ -4142,7 +4213,7 @@ namespace OpenTK.Windowing.GraphicsLibraryFramework
             Window* window,
             GLFWCallbacks.CursorPosCallback callback)
         {
-            return glfwSetCursorPosCallback(window, Marshal.GetFunctionPointerForDelegate(callback));
+            return glfwSetCursorPosCallback(window, callback);
         }
 
         /// <summary>
@@ -4173,7 +4244,7 @@ namespace OpenTK.Windowing.GraphicsLibraryFramework
             Window* window,
             GLFWCallbacks.DropCallback callback)
         {
-            return glfwSetDropCallback(window, Marshal.GetFunctionPointerForDelegate(callback));
+            return glfwSetDropCallback(window, callback);
         }
 
         /// <summary>
@@ -4206,7 +4277,7 @@ namespace OpenTK.Windowing.GraphicsLibraryFramework
         /// </remarks>
         public static IntPtr SetErrorCallback(GLFWCallbacks.ErrorCallback callback)
         {
-            return glfwSetErrorCallback(Marshal.GetFunctionPointerForDelegate(callback));
+            return glfwSetErrorCallback(callback);
         }
 
         /// <summary>
@@ -4330,7 +4401,7 @@ namespace OpenTK.Windowing.GraphicsLibraryFramework
         /// </remarks>
         public static IntPtr SetJoystickCallback(GLFWCallbacks.JoystickCallback callback)
         {
-            return glfwSetJoystickCallback(Marshal.GetFunctionPointerForDelegate(callback));
+            return glfwSetJoystickCallback(callback);
         }
 
         /// <summary>
@@ -4375,7 +4446,7 @@ namespace OpenTK.Windowing.GraphicsLibraryFramework
             Window* window,
             GLFWCallbacks.KeyCallback callback)
         {
-            return glfwSetKeyCallback(window, Marshal.GetFunctionPointerForDelegate(callback));
+            return glfwSetKeyCallback(window, callback);
         }
 
         /// <summary>
@@ -4404,7 +4475,7 @@ namespace OpenTK.Windowing.GraphicsLibraryFramework
             Window* window,
             GLFWCallbacks.ScrollCallback callback)
         {
-            return glfwSetScrollCallback(window, Marshal.GetFunctionPointerForDelegate(callback));
+            return glfwSetScrollCallback(window, callback);
         }
 
         /// <summary>
@@ -4427,7 +4498,7 @@ namespace OpenTK.Windowing.GraphicsLibraryFramework
         /// </remarks>
         public static IntPtr SetMonitorCallback(GLFWCallbacks.MonitorCallback callback)
         {
-            return glfwSetMonitorCallback(Marshal.GetFunctionPointerForDelegate(callback));
+            return glfwSetMonitorCallback(callback);
         }
 
         /// <summary>
@@ -4460,7 +4531,7 @@ namespace OpenTK.Windowing.GraphicsLibraryFramework
             Window* window,
             GLFWCallbacks.MouseButtonCallback callback)
         {
-            return glfwSetMouseButtonCallback(window, Marshal.GetFunctionPointerForDelegate(callback));
+            return glfwSetMouseButtonCallback(window, callback);
         }
 
         /// <summary>
@@ -4497,7 +4568,7 @@ namespace OpenTK.Windowing.GraphicsLibraryFramework
             Window* window,
             GLFWCallbacks.WindowCloseCallback callback)
         {
-            return glfwSetWindowCloseCallback(window, Marshal.GetFunctionPointerForDelegate(callback));
+            return glfwSetWindowCloseCallback(window, callback);
         }
 
         /// <summary>
@@ -4528,7 +4599,7 @@ namespace OpenTK.Windowing.GraphicsLibraryFramework
             Window* window,
             GLFWCallbacks.WindowFocusCallback callback)
         {
-            return glfwSetWindowFocusCallback(window, Marshal.GetFunctionPointerForDelegate(callback));
+            return glfwSetWindowFocusCallback(window, callback);
         }
 
         /// <summary>
@@ -4628,7 +4699,7 @@ namespace OpenTK.Windowing.GraphicsLibraryFramework
             Window* window,
             GLFWCallbacks.WindowIconifyCallback callback)
         {
-            return glfwSetWindowIconifyCallback(window, Marshal.GetFunctionPointerForDelegate(callback));
+            return glfwSetWindowIconifyCallback(window, callback);
         }
 
         /// <summary>
@@ -4745,7 +4816,7 @@ namespace OpenTK.Windowing.GraphicsLibraryFramework
             Window* window,
             GLFWCallbacks.WindowPosCallback callback)
         {
-            return glfwSetWindowPosCallback(window, Marshal.GetFunctionPointerForDelegate(callback));
+            return glfwSetWindowPosCallback(window, callback);
         }
 
         /// <summary>
@@ -4780,7 +4851,7 @@ namespace OpenTK.Windowing.GraphicsLibraryFramework
             Window* window,
             GLFWCallbacks.WindowRefreshCallback callback)
         {
-            return glfwSetWindowRefreshCallback(window, Marshal.GetFunctionPointerForDelegate(callback));
+            return glfwSetWindowRefreshCallback(window, callback);
         }
 
         /// <summary>
@@ -4846,7 +4917,7 @@ namespace OpenTK.Windowing.GraphicsLibraryFramework
             Window* window,
             GLFWCallbacks.WindowSizeCallback callback)
         {
-            return glfwSetWindowSizeCallback(window, Marshal.GetFunctionPointerForDelegate(callback));
+            return glfwSetWindowSizeCallback(window, callback);
         }
 
         /// <summary>
@@ -5683,9 +5754,9 @@ namespace OpenTK.Windowing.GraphicsLibraryFramework
             VkHandle instance,
             Window* window,
             void* allocator,
-            VkHandle surface)
+            out VkHandle surface)
         {
-            return glfwCreateWindowSurface(instance, window, allocator, surface);
+            return glfwCreateWindowSurface(instance, window, allocator, out surface);
         }
     }
 }